pyproject.toml
[project]
name = "rag-evaluation-framework"
version = "0.1.0"
description = "RAG Evaluation Framework with ChromaDB vector storage"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "chromadb>=1.0.0,<1.4.0",
    "llama-index>=0.9.0",
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-llms-openai>=0.1.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "streamlit>=1.28.0",
    "python-multipart>=0.0.21",
    "llama-index-vector-stores-chroma>=0.5.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
rag-cli = "scripts.cli:app"
rag-api = "api.server:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "eval", "rag_app", "scripts", "tests", "ui"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | storage
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "storage",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = [
    "storage/",
    "__pycache__/",
]
